Moved router into its own package, and made the endpoint for the routes programmatic
